Set default permalink for new posts going forward
At first I was confused because this change only impacts the two
posts from 2017 — all the prior posts for some reason were still using
what seemed to be the default permalinks of e.g. /post/:title — but
eventually I figured out that the reason for that was that all the prior
posts seem to have the property `url` set in their front matter, which
must have happened during some prior migration. So that means that these
URLs are what’s currently live — which I confirmed. So I’ll just leave
those older posts (the vast majority of them) alone and it’s all good.